requirements.txt
altgraph==0.17.4
certifi==2024.12.14
charset-normalizer==3.4.1
customtkinter==5.2.1
idna==3.10
mss==9.0.2
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pyasn1==0.6.1
pathgenerator
pynput>=1.7.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
requests==2.32.3
discord.py>=2.3.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
urllib3==2.3.0
Pillow>=8.0.1
PyQt5>=5.15.0
